refactor(protocol): slim suspension declaration graph
