+++ packages/gateway-protocol/src/schema/error-codes.ts
@@ -1,5 +1,5 @@
 // Gateway Protocol schema module defines protocol validation shapes.
-import type { ErrorShape } from "./types.js";
+import type { ErrorShape } from "./frames.js";
 
 /** Gateway JSON-RPC style error codes shared by clients and server handlers. */
 export const ErrorCodes = {
