+++ packages/gateway-protocol/src/schema/protocol-schemas.ts
@@ -223,7 +223,6 @@ import {
 } from "./frames.js";
 import {
   GatewaySuspendBlockerSchema,
-  GatewaySuspendCountsSchema,
   GatewaySuspendPrepareBusyResultSchema,
   GatewaySuspendPrepareParamsSchema,
   GatewaySuspendPrepareReadyResultSchema,
@@ -435,7 +434,6 @@ export const ProtocolSchemas = {
   StateVersion: StateVersionSchema,
   Snapshot: SnapshotSchema,
   ErrorShape: ErrorShapeSchema,
-  GatewaySuspendCounts: GatewaySuspendCountsSchema,
   GatewaySuspendTaskBlocker: GatewaySuspendTaskBlockerSchema,
   GatewaySuspendBlocker: GatewaySuspendBlockerSchema,
   GatewaySuspendPrepareParams: GatewaySuspendPrepareParamsSchema,
